fix: move components utils and data into core
