--- src/components/footer.astro
@@ -1,56 +1,0 @@
----
-import { cn } from "@/lib/utils";
-
-const linkClass = cn(
-  "transition-colors",
-  "text-foreground/70",
-  "hover:text-foreground"
-);
-
-const links = [
-  {
-    external: true,
-    href: "https://www.youtube.com/@alexandretrotel",
-    label: "YouTube",
-  },
-  {
-    external: true,
-    href: "https://www.github.com/alexandretrotel",
-    label: "GitHub",
-  },
-  {
-    external: true,
-    href: "/rss.xml",
-    label: "RSS",
-  },
-];
----
-
-<footer class="w-full bg-background">
-  <div
-    class="mx-auto flex h-16 max-w-3xl items-center justify-between text-sm text-muted-foreground"
-  >
-    <div />
-
-    <nav class="flex items-center gap-4">
-      {links.map((link, index) => (
-        <>
-          {index > 0 && (
-            <span aria-hidden="true" class="text-foreground/40">
-              /
-            </span>
-          )}
-
-          <a
-            class={linkClass}
-            href={link.href}
-            rel={link.external ? "noreferrer" : undefined}
-            target={link.external ? "_blank" : undefined}
-          >
-            {link.label}
-          </a>
-        </>
-      ))}
-    </nav>
-  </div>
-</footer>
