--- src/components/pages/home.astro
@@ -1,95 +1,0 @@
----
-import { getCollection } from "astro:content";
-import Hero from "@/components/pages/home/hero.astro";
-import Sections from "@/components/pages/home/sections.astro";
-import Layout from "@/layouts/main.astro";
-
-const metadata = {
-  description:
-    "Alexandre Trotel is a tech entrepreneur and full-stack developer.",
-  title: "Alexandre Trotel",
-};
-
-const blogHref = "/blog";
-
-const formatDate = (date: Date): string =>
-  date.toLocaleDateString("en-US", {
-    day: "numeric",
-    month: "short",
-    year: "numeric",
-  });
-
-const openSourceProjects = [
-  {
-    name: "todo-tree",
-    summary: "find and display TODO-style comments in your codebase.",
-    url: "https://github.com/alexandretrotel/todo-tree",
-  },
-  {
-    name: "mntn",
-    summary: "manage your dotfiles with the use of profiles.",
-    url: "https://github.com/alexandretrotel/mntn",
-  },
-  {
-    name: "feedyourai",
-    summary:
-      "combine multiple files into a single file for LLM processing with filtering options.",
-    url: "https://github.com/alexandretrotel/feedyourai",
-  },
-];
-
-const ventures = [
-  {
-    name: "Stealth Startup",
-    summary: "developing a next-generation product.",
-    url: "https://en.wikipedia.org/wiki/Stealth_startup",
-  },
-  {
-    name: "Zap Studio",
-    summary: "open source libraries, templates, and specifications.",
-    url: "https://github.com/zap-studio",
-  },
-];
-
-let latestPost: {
-  title: string;
-  url: string;
-  date: Date;
-} | null = null;
-
-try {
-  const posts = await getCollection("blog");
-  const publishedPosts = posts
-    .filter((p) => p.data.published !== false)
-    .toSorted((a, b) => b.data.date.getTime() - a.data.date.getTime());
-
-  const [p] = publishedPosts;
-  if (p) {
-    latestPost = {
-      date: p.data.date,
-      title: p.data.title,
-      url: `/blog/${p.id}/`,
-    };
-  }
-} catch {
-  latestPost = null;
-}
----
-
-<Layout title={metadata.title} description={metadata.description}>
-  <section class="mx-auto flex h-full w-full max-w-3xl flex-col gap-12 py-2">
-    <Hero />
-    <Sections
-      blogHref={blogHref}
-      latestBlogItem={latestPost
-        ? {
-            name: latestPost.title,
-            summary: formatDate(latestPost.date),
-            url: latestPost.url,
-          }
-        : null}
-      openSourceProjects={openSourceProjects}
-      ventures={ventures}
-    />
-  </section>
-</Layout>
