--- src/components/pages/home/blog-preview.astro
@@ -1,31 +1,0 @@
----
-import HomeSection from "@/components/ui/home-section.astro";
-import ItemList from "@/components/ui/item-list.astro";
-
-interface LatestBlogItem {
-  name: string;
-  summary: string;
-  url: string;
-}
-
-interface Props {
-  blogHref: string;
-  latestBlogItem: LatestBlogItem;
-}
-
-const props = Astro.props as Props;
-const { blogHref, latestBlogItem } = props;
----
-
-<HomeSection title="Blog">
-  <ItemList items={[latestBlogItem]} linkTarget="_self" />
-
-  <div class="pt-3">
-    <a
-      class="text-muted-foreground text-sm transition-colors hover:text-foreground"
-      href={blogHref}
-    >
-      Read all posts →
-    </a>
-  </div>
-</HomeSection>
