--- src/components/pages/home/sections.astro
@@ -1,39 +1,0 @@
----
-import BlogPreview from "@/components/pages/home/blog-preview.astro";
-import HomeSection from "@/components/pages/home/home-section.astro";
-import ItemList from "@/components/pages/home/item-list.astro";
-
-interface ListItem {
-  name: string;
-  summary: string;
-  url?: string;
-}
-
-interface LatestBlogItem {
-  name: string;
-  summary: string;
-  url: string;
-}
-
-interface Props {
-  blogHref: string;
-  latestBlogItem: LatestBlogItem | null;
-  openSourceProjects: ListItem[];
-  ventures: ListItem[];
-}
-
-const props = Astro.props as Props;
-const { blogHref, latestBlogItem, openSourceProjects, ventures } = props;
----
-
-<div class="space-y-10">
-  <HomeSection title="Open Source">
-    <ItemList items={openSourceProjects} />
-  </HomeSection>
-
-  <HomeSection title="Ventures">
-    <ItemList items={ventures} linkTarget="_blank" linkRel="noreferrer" />
-  </HomeSection>
-
-  {latestBlogItem && <BlogPreview blogHref={blogHref} latestBlogItem={latestBlogItem} />}
-</div>
