--- src/components/pages/not-found.astro
@@ -1,6 +1,0 @@
----
-import NotFoundContent from "@/components/pages/not-found/content.astro";
-import Layout from "@/layouts/main.astro";
----
-
-<Layout title="Not Found | Alexandre Trotel"> <NotFoundContent /> </Layout>
