--- src/components/pages/not-found/content.astro
@@ -1,12 +1,0 @@
----
-import Button from "@/components/ui/button.astro";
----
-
-<div class="flex h-full flex-col items-center justify-center gap-4 py-32">
-  <div class="flex flex-col items-center justify-center">
-    <p class="text-muted-foreground text-base">This page does not exist.</p>
-  </div>
-  <Button variant="ghost" class="active:scale-90 active:bg-accent">
-    <a href="/" class="text-sm">Go back</a>
-  </Button>
-</div>
