--- src/components/ui/button.astro
@@ -1,54 +1,0 @@
----
-import type { VariantProps } from "class-variance-authority";
-import { cva } from "class-variance-authority";
-import { cn } from "@/lib/utils";
-
-const buttonVariants = cva(
-  "inline-flex shrink-0 items-center justify-center gap-2 whitespace-nowrap rounded-md font-medium text-sm outline-none transition-all focus-visible:border-ring focus-visible:ring-[3px] focus-visible:ring-ring/50 disabled:pointer-events-none disabled:opacity-50 aria-invalid:border-destructive aria-invalid:ring-destructive/20 dark:aria-invalid:ring-destructive/40 [&_svg:not([class*='size-'])]:size-4 [&_svg]:pointer-events-none [&_svg]:shrink-0",
-  {
-    defaultVariants: {
-      size: "default",
-      variant: "default",
-    },
-    variants: {
-      size: {
-        default: "h-9 px-4 py-2 has-[>svg]:px-3",
-        icon: "size-9",
-        lg: "h-10 rounded-md px-6 has-[>svg]:px-4",
-        sm: "h-8 gap-1.5 rounded-md px-3 has-[>svg]:px-2.5",
-      },
-      variant: {
-        default:
-          "bg-primary text-primary-foreground shadow-xs hover:bg-primary/90",
-        destructive:
-          "bg-destructive text-white shadow-xs hover:bg-destructive/90 focus-visible:ring-destructive/20 dark:bg-destructive/60 dark:focus-visible:ring-destructive/40",
-        ghost:
-          "hover:bg-accent hover:text-accent-foreground dark:hover:bg-accent/50",
-        link: "text-primary underline-offset-4 hover:underline",
-        outline:
-          "border bg-background shadow-xs hover:bg-accent hover:text-accent-foreground dark:border-input dark:bg-input/30 dark:hover:bg-input/50",
-        secondary:
-          "bg-secondary text-secondary-foreground shadow-xs hover:bg-secondary/80",
-      },
-    },
-  }
-);
-
-type ButtonVariants = VariantProps<typeof buttonVariants>;
-
-interface Props {
-  class?: string;
-  size?: ButtonVariants["size"];
-  variant?: ButtonVariants["variant"];
-}
-
-const props = Astro.props as Props;
-const { variant, size, class: className } = props;
----
-
-<span
-  class={cn(buttonVariants({ variant, size, className }))}
-  data-slot="button"
->
-  <slot />
-</span>
