+++ src/core/components/components/dialog.astro
