+++ src/core/components/layout/header.astro
@@ -1,0 +1,26 @@
+---
+import DarkModeToggle from "@/core/components/components/dark-mode-toggle.svelte";
+---
+
+<div
+  class="top-0 z-50 h-16 w-full border-border/40 bg-background/95 backdrop-blur supports-backdrop-filter:bg-background/60"
+>
+  <div class="mx-auto flex h-14 max-w-3xl items-center justify-between py-8">
+    <div class="flex items-center gap-6">
+      <a
+        class="flex items-center gap-4 transition-transform active:scale-90"
+        href="/"
+      >
+        <img
+          alt="Logo"
+          class="rounded-md hover:opacity-80"
+          height="32"
+          width="32"
+          src="/logo.png"
+        >
+      </a>
+    </div>
+
+    <DarkModeToggle client :load />
+  </div>
+</div>
