+++ src/core/features/pages/blog-post/components/post-body.astro
