+++ src/core/features/pages/blog/components/header.astro
