+++ src/core/features/pages/home/components/hero.astro
