+++ src/core/features/pages/home/components/home-section.astro
