+++ src/core/features/pages/not-found/page.astro
@@ -1,0 +1,6 @@
+---
+import Layout from "@/core/components/layout/main-layout.astro";
+import NotFoundContent from "@/core/features/pages/not-found/components/content.astro";
+---
+
+<Layout title="Not Found | Alexandre Trotel"> <NotFoundContent /> </Layout>
