--- src/data/metadata.ts
@@ -1,6 +1,0 @@
-export const METADATA = {
-  BASE_URL: "https://www.alexandretrotel.org",
-  DESCRIPTION:
-    "Alexandre Trotel is a tech entrepreneur and full-stack developer.",
-  TITLE: "Alexandre Trotel",
-} as const;
