--- src/layouts/main.astro
@@ -1,106 +1,0 @@
----
-import "@/styles/globals.css";
-import "@fontsource-variable/geist";
-import "@fontsource-variable/lora";
-
-import { ClientRouter } from "astro:transitions";
-import Footer from "@/components/footer.astro";
-import Header from "@/components/header.astro";
-
-interface Props {
-  description?: string;
-  title?: string;
-}
-
-const defaultMetadata = {
-  description:
-    "Alexandre Trotel is a tech entrepreneur and full-stack developer.",
-  title: "Alexandre Trotel",
-};
-
-const props = Astro.props as Props;
-const {
-  title = defaultMetadata.title,
-  description = defaultMetadata.description,
-} = props;
-
-const ogLocale = "en_US";
----
-
-<!doctype html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta name="description" content={description}>
-    <meta name="generator" content={Astro.generator}>
-
-    <!-- Open Graph -->
-    <meta property="og:type" content="website">
-    <meta property="og:locale" content={ogLocale}>
-    <meta property="og:url" content={Astro.url.href}>
-    <meta property="og:title" content={title}>
-    <meta property="og:description" content={description}>
-
-    <!-- Twitter -->
-    <meta name="twitter:title" content={title}>
-    <meta name="twitter:description" content={description}>
-
-    <link rel="icon" type="image/png" href="/logo.png">
-    <title>{title}</title>
-
-    <!-- Sitemap -->
-    <link rel="sitemap" type="application/xml" href="/sitemap-index.xml">
-
-    <!-- RSS -->
-    {Astro.site && (
-      <link
-        href={new URL("rss.xml", Astro.site)}
-        rel="alternate"
-        title="Alexandre Trotel"
-        type="application/rss+xml"
-      />
-    )}
-
-    <!-- Katex -->
-    <link
-      rel="stylesheet"
-      href="https://cdn.jsdelivr.net/npm/katex@0.13.0/dist/katex.min.css"
-    >
-
-    <!-- Theme Script (prevents flash) -->
-    <script is:inline>
-      /* eslint-disable jest/require-hook */
-      const getThemePreference = () => {
-        if (
-          typeof localStorage !== "undefined" &&
-          localStorage.getItem("theme")
-        ) {
-          return localStorage.getItem("theme");
-        }
-        return window.matchMedia("(prefers-color-scheme: dark)").matches
-          ? "dark"
-          : "light";
-      };
-
-      const applyTheme = () => {
-        const theme = getThemePreference();
-        document.documentElement.classList.toggle("dark", theme === "dark");
-      };
-
-      document.addEventListener("astro:after-swap", applyTheme);
-      applyTheme();
-    </script>
-
-    <ClientRouter />
-  </head>
-  <body class="antialiased">
-    <div class="min-h-screen bg-background px-6">
-      <Header />
-      <main class="mx-auto min-h-[calc(100vh-8rem)] max-w-3xl py-12 md:py-18">
-        <slot />
-      </main>
-      <Footer />
-    </div>
-  </body>
-</html>
