+++ src/pages/404.astro
@@ -1,5 +1,5 @@
 ---
-import NotFoundPage from "@/components/pages/not-found.astro";
+import NotFoundPage from "@/core/features/pages/not-found/page.astro";
 
 export const prerender = true;
 ---
