+++ src/pages/blog/[slug].astro
@@ -1,8 +1,8 @@
 ---
 import type { CollectionEntry } from "astro:content";
 import { getCollection } from "astro:content";
-import BlogPostPage from "@/components/pages/blog-post.astro";
 import type { BlogCollection } from "@/content.config";
+import BlogPostPage from "@/core/features/pages/blog-post/page.astro";
 
 export const getStaticPaths = async () => {
   const posts: CollectionEntry<BlogCollection>[] = await getCollection("blog");
