+++ src/pages/blog/index.astro
@@ -1,5 +1,5 @@
 ---
-import BlogPage from "@/components/pages/blog.astro";
+import BlogPage from "@/core/features/pages/blog/page.astro";
 
 export const prerender = true;
 ---
