+++ src/pages/goals.astro
@@ -1,5 +1,5 @@
 ---
-import GoalsPage from "@/components/pages/goals.astro";
+import GoalsPage from "@/core/features/pages/goals/page.astro";
 
 export const prerender = true;
 ---
