+++ src/pages/index.astro
@@ -1,5 +1,5 @@
 ---
-import HomePage from "@/components/pages/home.astro";
+import HomePage from "@/core/features/pages/home/page.astro";
 
 export const prerender = true;
 ---
