+++ src/pages/rss.xml.ts
@@ -2,8 +2,8 @@ import type { CollectionEntry } from "astro:content";
 import { getCollection } from "astro:content";
 import rss from "@astrojs/rss";
 import type { APIContext } from "astro";
-
 import type { BlogCollection } from "@/content.config";
+import { SITE_CONFIG } from "@/shared/config/site";
 
 export const GET = async (context: APIContext) => {
   const blogPosts: CollectionEntry<BlogCollection>[] =
@@ -11,15 +11,14 @@ export const GET = async (context: APIContext) => {
 
   return rss({
     customData: "<language>en-us</language>",
-    description:
-      "Alexandre Trotel is a tech entrepreneur and full-stack developer.",
+    description: SITE_CONFIG.description,
     items: blogPosts.map((post) => ({
       description: post.data.description,
       link: `/blog/${post.id}/`,
       pubDate: post.data.date,
       title: post.data.title,
     })),
-    site: context.site ?? new URL("https://www.alexandretrotel.org"),
-    title: "Alexandre Trotel",
+    site: context.site ?? new URL(SITE_CONFIG.baseUrl),
+    title: SITE_CONFIG.title,
   });
 };
