+++ src/shared/config/site.ts
@@ -1,0 +1,6 @@
+export const SITE_CONFIG = {
+  baseUrl: "https://www.alexandretrotel.org",
+  description:
+    "Alexandre Trotel is a tech entrepreneur and full-stack developer.",
+  title: "Alexandre Trotel",
+} as const;
